Update UC Santa Cruz stipend:
This assumes student is a 50% GSR during the school year and 100% GSR in the summer:

Pre-ATC is step 1:
5090/month @ 100%
(5090/2)*9 + 5090*3 = 38175 / year

Post-ATC is step 3:
5909/month @ 100%
(5909/2)*9 + 5909*3 = 44317 / year

We also offer a $2500 housing stipend to all graduate students. So,

pre-ATC: 40675
post-ATC: 46817

Salaries updated as of 4/1/23 salary table:
https://apo.ucsc.edu/docs/scales-crnt.pdf

I'm assuming that students get summer support since other institutions
are including it. I filed another issue to discuss these methodologies.